Cliente service e salva os dados do cliente no json

+++ src/main/java/dados/IRepositorioCliente.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface IRepositorioCliente {
     void cadastrar(Cliente cliente);
-    Cliente buscar(int id);
+    Cliente buscar(long id);
     Cliente buscarPorEmail(String email);
     List<Cliente> listar();
     void atualizar(Cliente cliente);
-    void remover(int id);
+    void remover(long id);
 }
